// autogpt_platform/frontend/src/app/api/__generated__/models/agentExecutionStatus.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type AgentExecutionStatus =
  (typeof AgentExecutionStatus)[keyof typeof AgentExecutionStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AgentExecutionStatus = {
  INCOMPLETE: "INCOMPLETE",
  QUEUED: "QUEUED",
  RUNNING: "RUNNING",
  COMPLETED: "COMPLETED",
  TERMINATED: "TERMINATED",
  FAILED: "FAILED",
} as const;