// autogpt_platform/frontend/src/app/api/__generated__/models/graphMeta.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { GraphMetaRecommendedScheduleCron } from "./graphMetaRecommendedScheduleCron";
import type { GraphMetaForkedFromId } from "./graphMetaForkedFromId";
import type { GraphMetaForkedFromVersion } from "./graphMetaForkedFromVersion";
import type { BaseGraphOutput } from "./baseGraphOutput";
import type { GraphMetaInputSchema } from "./graphMetaInputSchema";
import type { GraphMetaOutputSchema } from "./graphMetaOutputSchema";
import type { GraphMetaTriggerSetupInfo } from "./graphMetaTriggerSetupInfo";
import type { GraphMetaCredentialsInputSchema } from "./graphMetaCredentialsInputSchema";

export interface GraphMeta {
  id?: string;
  version?: number;
  is_active?: boolean;
  name: string;
  description: string;
  recommended_schedule_cron?: GraphMetaRecommendedScheduleCron;
  forked_from_id?: GraphMetaForkedFromId;
  forked_from_version?: GraphMetaForkedFromVersion;
  sub_graphs?: BaseGraphOutput[];
  user_id: string;
  readonly input_schema: GraphMetaInputSchema;
  readonly output_schema: GraphMetaOutputSchema;
  readonly has_external_trigger: boolean;
  readonly trigger_setup_info: GraphMetaTriggerSetupInfo;
  readonly credentials_input_schema: GraphMetaCredentialsInputSchema;
}